refactor: rename arcgis-rest-feature-service to arcgis-rest-feature-layer
AFFECTS PACKAGES:
@esri/arcgis-rest-feature-layer
@esri/arcgis-rest-geocoding
@esri/arcgis-rest-portal
@esri/arcgis-rest-request
