+++ packages/arcgis-rest-feature-layer/src/decodeValues.ts
@@ -44,7 +44,7 @@ export interface IDecodeValuesRequestOptions extends IRequestOptions {
 
 /**
  * ```js
- * import { queryFeatures, decodeValues } from '@esri/arcgis-rest-feature-service';
+ * import { queryFeatures, decodeValues } from '@esri/arcgis-rest-feature-layer';
  * //
  * const url = `https://sampleserver6.arcgisonline.com/arcgis/rest/services/ServiceRequest/FeatureServer/0`
  * queryFeatures({ url })
