+++ packages/arcgis-rest-feature-layer/src/helpers.ts
