+++ packages/arcgis-rest-feature-layer/src/query.ts
@@ -109,7 +109,7 @@ export interface IQueryResponse {
 
 /**
  * ```js
- * import { getFeature } from '@esri/arcgis-rest-feature-service';
+ * import { getFeature } from '@esri/arcgis-rest-feature-layer';
  * //
  * const url = "https://services.arcgis.com/V6ZHFr6zdgNZuVG0/arcgis/rest/services/Landscape_Trees/FeatureServer/0";
  * //
@@ -145,7 +145,7 @@ export function getFeature(
 
 /**
  * ```js
- * import { queryFeatures } from '@esri/arcgis-rest-feature-service';
+ * import { queryFeatures } from '@esri/arcgis-rest-feature-layer';
  * //
  * queryFeatures({
  *   url: "http://sampleserver6.arcgisonline.com/arcgis/rest/services/Census/MapServer/3",
