+++ packages/arcgis-rest-feature-layer/test/mocks/cvdQueryResponse.ts
