+++ packages/arcgis-rest-feature-layer/test/mocks/feature.ts
