+++ packages/arcgis-rest-feature-layer/test/mocks/fields.ts
