+++ packages/arcgis-rest-feature-layer/test/mocks/foo.txt
