+++ packages/arcgis-rest-feature-layer/test/mocks/service.ts
