+++ packages/arcgis-rest-feature-layer/tsconfig.json
