+++ packages/arcgis-rest-geocoding/README.md
@@ -19,7 +19,6 @@
 ### Example
 
 ```bash
-npm install @esri/arcgis-rest-common-types
 npm install @esri/arcgis-rest-request
 npm install @esri/arcgis-rest-geocoding
 ```
